Transition 64 bit: warnings deleting #2

+++ PdfReportKit/PRKGenerator.m
@@ -214,7 +214,7 @@ static NSArray * reportDefaultTags = nil;
     else if (sectionType == PRKSectionTypeFooter)
         footerFormatter = formatter;
     else
-        [NSException raise:@"Invalid Section Type" format:@"Section Type: %ld is invalid", sectionType];
+        [NSException raise:@"Invalid Section Type" format:@"Section Type: %ld is invalid", (long)sectionType];
 }
 
 - (void)renderPage
